refactor: took away code duplication
