+++ app/shared/components/accordion-blocks/editable-section-list/EditableSectionList.test.tsx
